requirements.txt
allure-behave
behave
behave-webdriver
pyyaml
selenium
unittest2
assertpy
